Add failing test to show hex overflow
When an RGB color with values below zero or greater than 1 is converted
to hex, the resulting hex string overflows. A value greater than 1
should always be rendered as FF, while a value below zero should always
be 00.

These RGB colors can result from LUV colors outside the RGB space being
converted to RGB. When such a color is rendered as hex, the result is
wrong, as this test illustrates.