Fix hex overflow on RGB values outside of [0,1]
When an RGB color with values either above 1 or below zero is converted
to hex, the unit8 overflows and wrap around.

Values below 0 become FF and decreasing,
Values above 1 become 00 and increasing.

These RGB values can result from LUV colors outside the RGB space being
converted into RGB. To preseve the LUV space, apparently it was decided
that RGB values outside of [0,1] are acceptable (to represent the RGB
color without loss), but especially when rendering to hex these
overflowing values become a problem.

A color "more red" than 100% RGB red should not be black, but RGB 100%
red, instead.
A yellow that is just a little bit more red than FFFF00 should not
be green (00FF00) suddenly, but instead the best yellow we can get out
of RGB, which is FFFF00.

-> Clamping the RGB values before hex conversion solves this without losing
the information in the Color object.